+++ tiempo/contrib/django_app/templates/tiempo/recent.html
